backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
docling
docling-core
neo4j
sentence-transformers
transformers
numpy<2
python-dotenv
pydantic